Fixes #1290, Fixes #1467: Load country specific-content, and fix country list bug
By default we will use locale-country, but once a user selects a country,
we will actually use that country in the content YAML loading.

This starts the proper refactor of UserPreferences into a Store to use
it as an observable dependency.

Dependency: observer_producer - a verified dep I open-sourced from work