[WIP] Switch from MIT to Apache 2.0 License
Work in Progress. For discussion only at this point until we decide as a team to proceed. The Apache 2.0 is a stronger license than the MIT license along with having an explicit patent grant for code contributors. It's also the preferred license of the Linux Foundation. We've been approached by the Linux Foundation for Public Health who are interested in supporting this project. You can read more about the license differences here: https://snyk.io/blog/mit-apache-bsd-fairest-of-them-all/

IF AND WHEN WE AGREE....

Please everyone reply with "I Agree" to allow us to relicense the project under the Apache 2.0 license. It needs every one of the prior contributors to agree to this. At that point, the codebase will be licensed under two licenses. Going forward the plan is to use only the Apache 2.0 license.