StatsStore 'cache'
The StatsStore now tracks the country and time that the last stats loaded .
It now will not (re)load stats if the country is unset, or if the country is the same but the time has been too long.

Ideally, this system would be replaced with a single load and something like the RefreshIndicator class, but that is a much larger change.